requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.18
Flask-Bootstrap4==4.0.2
waitress==2.0.0
pandas==1.2.4
python-dotenv==0.17.1
psycopg2==2.8.6
invoke==1.5.0
